+++ lib/ruby_speech/grxml.rb
@@ -4,6 +4,8 @@ require 'ruby_speech/grxml/element'
 module RubySpeech
   module GRXML
     InvalidChildError = Class.new StandardError
+    MissingReferenceError = Class.new StandardError
+    ReferentialLoopError = Class.new StandardError
 
     GRXML_NAMESPACE = 'http://www.w3.org/2001/06/grammar'