+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # [develop](https://github.com/benlangfeld/ruby_speech)
   * Feature: Permit percentage rate values for prosody tags
+  * Bugfix: Rulerefs referenced n-levels deep under Rulerefs should be expanded.
+  * Bugfix: Optimize performance of built-in number DTMF grammar
 
 # [2.3.2](https://github.com/benlangfeld/ruby_speech/compare/v2.3.1...v2.3.2) - [2014-04-21](https://rubygems.org/gems/ruby_speech/versions/2.3.2)
   * Bugfix: String nodes should take non-strings and cast to a string (`#to_s`)